prometheus-node-exporter-lua: Export station metrics even if hostapd does not
Hostapd would only expose vht_caps_info for vht capable radios, and that
would result in the metric only be set in that case. But the basic truth
is that every station that joins an AP using a standard that does not
support VHT or beamforming also does not support it.

This way we get a consistent view over all connected stations.

Signed-off-by: Martin Weinelt <hexa@darmstadt.ccc.de>